Validation section
I solved the validation section (can't register without the given the required info). I tried to use HTML built in email validation but I couldn't implemented to the final check at the register button so I used the wide used email validation method -"An email is a string (a subset of ASCII characters)".